Split FrontEndTests out into five different parts
